# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ShopFlow Platform API
  description: |
    REST API for the wholesale/retail shop management platform (React web and React Native clients).

    * Authentication: mobile number + OTP → short-lived bearer access token (JWT) plus a rotating refresh token.
    * Authorization: role (OWNER, ADMIN, CUSTOMER) plus permission codes; customers can only access their own data.
    * Envelope: every JSON response is `{ success, data, message, pagination?, error?, requestId }`.
    * Errors: `error.code` is one of the documented business codes (e.g. INSUFFICIENT_STOCK, CREDIT_LIMIT_EXCEEDED).
    * Money: decimal numbers with 2 places, calculated only by the backend.
    * Lists: `page` (1-based), `pageSize` (max 100), `sort=field,asc|desc`.
    * Retries: send `Idempotency-Key` on order creation, payment recording, invoice generation and WhatsApp sends.
  version: v1
servers:
- url: /
  description: Current host
...
paths:
...
  /api/v1/business/logo:
    post:
      tags:
      - Business Settings
      summary: Upload the business logo
      description: "PNG, JPEG or WebP up to 5 MB."
      operationId: uploadLogo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessProfileResponse"
      security:
      - bearerAuth: []
  /api/v1/business/invoice-settings/signature:
    post:
      tags:
      - Business Settings
      summary: Upload the authorized-signature image
      operationId: uploadSignature
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceSettingsResponse"
      security:
      - bearerAuth: []
  /api/v1/business/bank-accounts:
    get:
      tags:
      - Business Settings
      summary: List bank accounts
      operationId: bankAccounts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListBankAccountResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Business Settings
      summary: Add a bank account
      description: The first account becomes the default printed on invoices.
      operationId: addBankAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBankAccountResponse"
      security:
      - bearerAuth: []
...
  /api/v1/my/profile:
    get:
      tags:
      - My Account
      summary: My business profile
      operationId: profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCustomerDetail"
      security:
      - bearerAuth: []
    patch:
      tags:
      - My Account
      summary: Update my business profile
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseCustomerDetail"
      security:
      - bearerAuth: []
...
  /api/v1/business:
    get:
      tags:
      - Business Settings
      summary: Get the business profile
      description: Available to every signed-in user (seller details shown on invoices).
      operationId: get_4
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessProfileResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Business Settings
      summary: Update the business profile
      description: Only provided fields change. State code is derived from the state
        name when omitted.
      operationId: update_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessProfileResponse"
      security:
      - bearerAuth: []
  /api/v1/business/tax-settings:
    get:
      tags:
      - Business Settings
      summary: Get tax settings
      description: Available to every signed-in user; needed to render tax information.
      operationId: taxSettings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaxSettingsResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Business Settings
      summary: Update tax settings
      operationId: updateTaxSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaxSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaxSettingsResponse"
      security:
      - bearerAuth: []
  /api/v1/business/settings:
    get:
      tags:
      - Business Settings
      summary: Get operational business rules
      description: "Credit policy, cancellation window, partial-delivery invoicing,\
        \ notifications, retention."
      operationId: settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessSettingsResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Business Settings
      summary: Update operational business rules
      operationId: updateSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBusinessSettingsResponse"
      security:
      - bearerAuth: []
  /api/v1/business/invoice-settings:
    get:
      tags:
      - Business Settings
      summary: Get invoice settings
      operationId: invoiceSettings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceSettingsResponse"
      security:
      - bearerAuth: []
    patch:
      tags:
      - Business Settings
      summary: Update invoice settings
      description: Numbering changes apply to sequences created for a new financial
        year; issued numbers never change.
      operationId: updateInvoiceSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceSettingsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseInvoiceSettingsResponse"
      security:
      - bearerAuth: []
  /api/v1/business/bank-accounts/{id}:
    patch:
      tags:
      - Business Settings
      summary: Update a bank account
      operationId: updateBankAccount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBankAccountRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBankAccountResponse"
      security:
      - bearerAuth: []
...
  /api/v1/business/states:
    get:
      tags:
      - Business Settings
      summary: List Indian states with GST state codes
      operationId: states
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringString"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldIssue"
...
    FieldIssue:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseMapStringString:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
    AddressResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
...
    ApiResponseCustomerDetail:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/CustomerDetail"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    CreditProfileResponse:
      type: object
      properties:
        creditEnabled:
          type: boolean
        creditLimit:
          type: number
        creditDays:
          type: integer
          format: int32
        creditPolicyOverride:
          type: string
          enum:
          - BLOCK
          - REQUIRE_ADMIN_APPROVAL
          - ALLOW
    CustomerDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerCode:
          type: string
        shopName:
          type: string
        contactName:
          type: string
        mobileNumber:
          type: string
        alternateMobile:
          type: string
        email:
          type: string
        gstin:
          type: string
        pan:
          type: string
        status:
          type: string
        statusReason:
          type: string
        statusChangedAt:
          type: string
          format: date-time
        notes:
          type: string
        hasLogin:
          type: boolean
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/AddressResponse"
        credit:
          $ref: "#/components/schemas/CreditProfileResponse"
        outstanding:
          $ref: "#/components/schemas/Outstanding"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Outstanding:
      type: object
      properties:
        ledgerBalance:
          type: number
        invoiceOutstanding:
          type: number
        overdueAmount:
          type: number
        openInvoiceCount:
          type: integer
          format: int32
        oldestDueDate:
          type: string
          format: date
        creditEnabled:
          type: boolean
        creditLimit:
          type: number
        creditDays:
          type: integer
          format: int32
        availableCredit:
          type: number
        creditPolicy:
          type: string
          enum:
          - BLOCK
          - REQUIRE_ADMIN_APPROVAL
          - ALLOW
...
    ApiResponseBusinessProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/BusinessProfileResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    BusinessProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        legalName:
          type: string
        logoFileId:
          type: string
          format: uuid
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        stateCode:
          type: string
        pincode:
          type: string
        phone:
          type: string
        mobile:
          type: string
        email:
          type: string
        gstin:
          type: string
        pan:
          type: string
        timezone:
          type: string
        currency:
          type: string
        financialYearStartMonth:
          type: integer
          format: int32
        termsAndConditions:
          type: string
        authorizedSignatory:
          type: string
    ApiResponseInvoiceSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/InvoiceSettingsResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    InvoiceSettingsResponse:
      type: object
      properties:
        invoicePrefix:
          type: string
        numberPadding:
          type: integer
          format: int32
        startingNumber:
          type: integer
          format: int64
        defaultPaymentTerms:
          type: string
        defaultTerms:
          type: string
        defaultFooter:
          type: string
        declaration:
          type: string
        signatureFileId:
          type: string
          format: uuid
        templateCode:
          type: string
        showBankDetails:
          type: boolean
    BankAccountRequest:
      type: object
      properties:
        bankName:
          type: string
          maxLength: 200
          minLength: 0
        accountName:
          type: string
          maxLength: 200
          minLength: 0
        accountNumber:
          type: string
          minLength: 1
          pattern: "^[0-9]{6,20}$"
        ifsc:
          type: string
          minLength: 1
          pattern: "^[A-Z]{4}0[A-Z0-9]{6}$"
        branch:
          type: string
          maxLength: 200
          minLength: 0
        isDefault:
          type: boolean
        active:
          type: boolean
      required:
      - accountNumber
      - bankName
      - ifsc
    ApiResponseBankAccountResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/BankAccountResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    BankAccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bankName:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        ifsc:
          type: string
        branch:
          type: string
        isDefault:
          type: boolean
        active:
          type: boolean
...
    UpdateCustomerRequest:
      type: object
      properties:
        shopName:
          type: string
          maxLength: 200
          minLength: 0
        contactName:
          type: string
          maxLength: 200
          minLength: 0
        gstin:
          type: string
          pattern: "^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z][1-9A-Z]Z[0-9A-Z]$"
        pan:
          type: string
          pattern: "^[A-Z]{5}[0-9]{4}[A-Z]$"
        email:
          type: string
          format: email
          maxLength: 200
          minLength: 0
        alternateMobile:
          type: string
        notes:
          type: string
          maxLength: 2000
          minLength: 0
...
    UpdateBusinessProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        legalName:
          type: string
          maxLength: 200
          minLength: 0
        addressLine1:
          type: string
          maxLength: 200
          minLength: 0
        addressLine2:
          type: string
          maxLength: 200
          minLength: 0
        city:
          type: string
          maxLength: 100
          minLength: 0
        state:
          type: string
          maxLength: 100
          minLength: 0
        stateCode:
          type: string
          pattern: "^[0-9]{2}$"
        pincode:
          type: string
          pattern: "^[1-9][0-9]{5}$"
        phone:
          type: string
          maxLength: 20
          minLength: 0
        mobile:
          type: string
          maxLength: 20
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 200
          minLength: 0
        gstin:
          type: string
          pattern: "^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z][1-9A-Z]Z[0-9A-Z]$"
        pan:
          type: string
          pattern: "^[A-Z]{5}[0-9]{4}[A-Z]$"
        financialYearStartMonth:
          type: integer
          format: int32
          maximum: 12
          minimum: 1
        termsAndConditions:
          type: string
          maxLength: 4000
          minLength: 0
        authorizedSignatory:
          type: string
          maxLength: 200
          minLength: 0
    UpdateTaxSettingsRequest:
      type: object
      properties:
        roundOffEnabled:
          type: boolean
        defaultGstRate:
          type: number
          maximum: 100
          minimum: 0
        allowedGstRates:
          type: array
          items:
            type: number
            maximum: 100
            minimum: 0
    ApiResponseTaxSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/TaxSettingsResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    TaxSettingsResponse:
      type: object
      properties:
        calculationMode:
          type: string
        roundOffEnabled:
          type: boolean
        defaultGstRate:
          type: number
        allowedGstRates:
          type: array
          items:
            type: number
        interStateTaxType:
          type: string
        intraStateTaxType:
          type: string
    UpdateBusinessSettingsRequest:
      type: object
      properties:
        creditPolicy:
          type: string
          enum:
          - BLOCK
          - REQUIRE_ADMIN_APPROVAL
          - ALLOW
        customerCancelAllowedUntil:
          type: string
          enum:
          - PLACED
          - ACCEPTED
          - NEVER
        showStockToCustomers:
          type: boolean
        gstinRequiredForCustomers:
          type: boolean
        panRequiredForCustomers:
          type: boolean
        partialDeliveryInvoicePolicy:
          type: string
          enum:
          - INVOICE_ACCEPTED_QUANTITY
          - INVOICE_DELIVERED_QUANTITY
        defaultCreditDays:
          type: integer
          format: int32
          maximum: 365
          minimum: 0
        defaultCreditLimit:
          type: number
          minimum: 0
        whatsappEnabled:
          type: boolean
        whatsappSender:
          type: string
          maxLength: 40
          minLength: 0
        whatsappInvoiceTemplate:
          type: string
          maxLength: 100
          minLength: 0
        notifyPush:
          type: boolean
        notifySms:
          type: boolean
        notifyEmail:
          type: boolean
        notifyWhatsapp:
          type: boolean
        dataRetentionYears:
          type: integer
          format: int32
          maximum: 50
          minimum: 1
    ApiResponseBusinessSettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/BusinessSettingsResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    BusinessSettingsResponse:
      type: object
      properties:
        creditPolicy:
          type: string
          enum:
          - BLOCK
          - REQUIRE_ADMIN_APPROVAL
          - ALLOW
        customerCancelAllowedUntil:
          type: string
          enum:
          - PLACED
          - ACCEPTED
          - NEVER
        showStockToCustomers:
          type: boolean
        gstinRequiredForCustomers:
          type: boolean
        panRequiredForCustomers:
          type: boolean
        partialDeliveryInvoicePolicy:
          type: string
          enum:
          - INVOICE_ACCEPTED_QUANTITY
          - INVOICE_DELIVERED_QUANTITY
        defaultCreditDays:
          type: integer
          format: int32
        defaultCreditLimit:
          type: number
        whatsappEnabled:
          type: boolean
        whatsappSender:
          type: string
        whatsappInvoiceTemplate:
          type: string
        notifyPush:
          type: boolean
        notifySms:
          type: boolean
        notifyEmail:
          type: boolean
        notifyWhatsapp:
          type: boolean
        dataRetentionYears:
          type: integer
          format: int32
    UpdateInvoiceSettingsRequest:
      type: object
      properties:
        invoicePrefix:
          type: string
          pattern: "^[A-Z0-9]{1,10}$"
        numberPadding:
          type: integer
          format: int32
          maximum: 10
          minimum: 3
        startingNumber:
          type: integer
          format: int64
          minimum: 1
        defaultPaymentTerms:
          type: string
          maxLength: 200
          minLength: 0
        defaultTerms:
          type: string
          maxLength: 4000
          minLength: 0
        defaultFooter:
          type: string
          maxLength: 500
          minLength: 0
        declaration:
          type: string
          maxLength: 2000
          minLength: 0
        templateCode:
          type: string
          pattern: "^[A-Z0-9_]{1,40}$"
        showBankDetails:
          type: boolean
    UpdateBankAccountRequest:
      type: object
      properties:
        bankName:
          type: string
          maxLength: 200
          minLength: 0
        accountName:
          type: string
          maxLength: 200
          minLength: 0
        accountNumber:
          type: string
          pattern: "^[0-9]{6,20}$"
        ifsc:
          type: string
          pattern: "^[A-Z]{4}0[A-Z0-9]{6}$"
        branch:
          type: string
          maxLength: 200
          minLength: 0
        isDefault:
          type: boolean
        active:
          type: boolean
...
    ApiResponseListBankAccountResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/BankAccountResponse"
        message:
          type: string
        pagination:
          $ref: "#/components/schemas/Pagination"
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT